Greation
Creating a dynamic greeting based on the time of day in React involves writing a small function that determines the current time and then displays an appropriate greeting (e.g., "Good Morning," "Good Afternoon," "Good Evening"). Below is a description of how you can implement this in a React component.
Steps to Create a Dynamic Greeting in React:
Setup a React Component:
Create a React functional component or class component where the greeting will be displayed.

Determine the Current Time:
Use JavaScript's Date object to get the current hour. Based on the hour, determine which greeting to display.

Create a Function for the Greeting:
Write a function that returns a different greeting depending on the current hour.

Render the Greeting:
In the render method (or JSX return), use the function to display the appropriate greeting.